# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Dental Backend API
  version: 1.0.0
  description: OpenAPI specification for dental clinic backend (services, appointments, payments, users, admin analytics, auth, chat)
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
paths:
...
  /api/appointments:
    post:
      summary: Create appointment (receptionist/admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAppointmentRequest'
      responses:
        '200':
          description: Created appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Appointment'
  /api/appointments/my:
    get:
      summary: Dentist's appointments
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AppointmentList'
  /api/appointments/history:
    get:
      summary: Customer appointment history
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Appointment history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AppointmentList'
  /api/appointments/{id}:
    put:
      summary: Update own appointment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
      responses:
        '200':
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Appointment'
    delete:
      summary: Cancel own appointment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
  /api/appointments/all:
    get:
      summary: List all appointments (admin)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All appointments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AppointmentList'
  /api/appointments/admin/{id}:
    put:
      summary: Admin update appointment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Appointment'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Appointment'
    delete:
      summary: Admin delete appointment
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
...
  /api/payments/appointment/{appointmentId}:
    get:
      summary: Get payments for appointment
      security:
        - bearerAuth: []
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Payments list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PaymentList'
...
  /api/admin/revenue/total:
    get:
      summary: Total revenue
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Total revenue amount
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Double'
  /api/admin/revenue/by-day:
    get:
      summary: Revenue by day
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Revenue map by date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_RevenueByDay'
  /api/admin/service-usage:
    get:
      summary: Service usage counts
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Service usage map
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_ServiceUsage'
  /api/auth/register:
    post:
      summary: Register new user
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: Registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_String'
  /api/auth/login:
    post:
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: JWT tokens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_JwtResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    ApiResponse_String:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: string
    ApiResponse_Double:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: number
              format: double
...
    Appointment:
      type: object
      properties:
        id:
          type: integer
        customerId:
          type: integer
        dentistId:
          type: integer
        receptionistId:
          type: integer
        status:
          type: string
        scheduledTime:
          type: string
          format: date-time
        notes:
          type: string
    ApiResponse_Appointment:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/Appointment'
    ApiResponse_AppointmentList:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Appointment'
    CreateAppointmentRequest:
      type: object
      required: [customerId, dentistId, scheduledTime]
      properties:
        customerId:
          type: integer
        dentistId:
          type: integer
        scheduledTime:
          type: string
          format: date-time
        notes:
          type: string
    Payment:
      type: object
      properties:
        id:
          type: integer
        appointmentId:
          type: integer
        amount:
          type: number
          format: double
        method:
          type: string
        invoiceNumber:
          type: string
        paidAt:
          type: string
          format: date-time
...
    ApiResponse_PaymentList:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Payment'
...
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        captchaToken:
          type: string
    LoginRequest:
      type: object
      properties:
        usernameOrEmail:
          type: string
        password:
          type: string
    JwtResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
    ApiResponse_JwtResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/JwtResponse'
...
    ApiResponse_RevenueByDay:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              additionalProperties:
                type: number
                format: double
    ApiResponse_ServiceUsage:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              additionalProperties:
                type: integer
...